requirements.txt
appdirs==1.4.4
attrs==21.2.0
bcrypt==4.1.3
beautifulsoup4==4.10.0
beniget==0.4.1
binwalk==2.3.3
blinker==1.4
Brotli==1.0.9
CacheControl==0.14.0
cachetools==5.5.0
capstone==5.0.1
certifi==2024.6.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
colored-traceback==0.3.0
command-not-found==0.3
crypto==1.4.1
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
distro==1.7.0
distro-info==1.1+ubuntu0.2
firebase-admin==6.5.0
fonttools==4.29.1
fs==2.4.12
gast==0.5.2
gmpy2==2.1.2
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.2
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.66.0
grpcio-status==1.66.0
html5lib==1.1
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
intervaltree==3.1.0
jeepney==0.7.1
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.8.0
lz4==3.1.3+dfsg
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.5.1
mercurial==6.1.1
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.8
Naked==0.1.32
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==3.4.0
Pillow==9.0.1
plumbum==1.8.3
ply==3.11
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
pwntools==4.12.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.21.0
pyelftools==0.31
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.9.0
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyqtgraph==0.12.4
pyserial==3.5
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-dotenv==1.0.1
pythran==0.10.0
pytz==2022.1
PyYAML==5.4.1
requests==2.32.3
ROPGadget==7.4
rpyc==6.0.0
rsa==4.9
scipy==1.8.0
SecretStorage==3.3.1
shellescape==3.8.1
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sympy==1.9
systemd-python==234
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
unicorn==2.0.1.post1
unix-ar==0.2.1
uritemplate==4.1.1
urllib3 @ file:///home/daiki/urllib3
wadllib==1.3.6
webencodings==0.5.1
zipp==1.0.0
zstandard==0.22.0
